SimpleIdentifier to remove strict Identifier name for ConfigFile

+++ src/main/java/dev/proststuff/utilitary/api/utility/Color.java
@@ -102,7 +102,7 @@ public class Color {
     }
 
     public int setRed(float r) {
-        return setRed(r * 255);
+        return setRed((int) r * 255);
     }
 
     public int setGreen(int g) {
@@ -111,7 +111,7 @@ public class Color {
     }
 
     public int setGreen(float g) {
-        return setGreen(g * 255);
+        return setGreen((int) g * 255);
     }
 
     public int setBlue(int b) {
@@ -120,7 +120,7 @@ public class Color {
     }
 
     public int setBlue(float b) {
-        return setBlue(b * 255);
+        return setBlue((int) b * 255);
     }
 
     @Override
